`Added .NET build and publish workflow in publish.yml`
This commit introduces a new workflow named `.NET` in the `publish.yml` file. The workflow is designed to build a .NET project and is triggered on release and workflow dispatch events. It runs on the latest version of Ubuntu and has read permissions for contents and write permissions for packages.

The workflow includes several steps such as setting up .NET, restoring dependencies, building the project, running tests, and publishing the project. It also includes steps for logging into the container registry, extracting metadata for Docker, and building and pushing the Docker image.

The final steps of the workflow involve initializing git, creating a new branch, copying files, committing the changes, and pushing the changes to the `community-operators` repository on GitHub using a secret token for authentication.